pyproject.toml
[tool.poetry]
name = "numpy-ros"
version = "0.1.3"
description = "Seamlessly convert between ROS messages and NumPy arrays."
authors = ["Lucas Wolf <iamlucaswolf@gmail.com>"]
keywords = ["numpy", "ros"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/iamlucaswolf/numpy-ros"
repository = "https://github.com/iamlucaswolf/numpy-ros"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.6, <3.10"
numpy = "^1.19.5"
numpy-quaternion = "^2021.4.5"
numba = {version = "^0.53.1", optional = true}
scipy = {version = ">=0.18.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
vermin = "^1.2.0"
coverage = {extras = ["toml"], version = "^5.5"}
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
quaternion-extras = ["numba", "scipy"]